pyproject.toml
[tool.poetry]
name = "videogames-system-cdk-py"
version = "1.0.0"
description = "Basic infrastructure for a project with AWS CDK and poetry"
authors = ["Ruben Moreno"]
readme = "README.md"
packages = [{ include = "videogames_system" }]

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
boto3 = "1.34.138"
botocore = "1.34.138"
aws-cdk-lib = "2.158.0"
constructs = "10.3.0"

[tool.poetry.group.dev.dependencies]
black = "24.4.2"
pre-commit = "3.7.1"
isort = "5.13.2"
boto3-stubs = { version = "1.34.135", extras = ["lambda", "s3"] }
pytest-cov = "5.0.0"
pytype = "2024.4.11"
pydocstyle = "6.3.0"
flake8 = "7.1.0"
pylint = "3.2.5"

[tool.poetry.scripts]
cdk-main-app = "videogames_system.aws_resources.cdk.index:main"

[tool.black]
line-length = 120
include = '.*\.py$'
exclude = '''
(
    \.git
  | __pycache__
  | \.venv
  | build
  | dist
  | setup.py
  | infrastructure
  | notebooks
)
'''

[tool.isort]
profile = 'black'
skip = [".git", "__pycache__", ".venv", "build", "dist", "setup.py"]

[tool.pydocstyle]
convention = "google"
add-ignore = "D100,D101,D102,D103,D104,D105,D106,D107"

[build-system]
requires = ["poetry>=1.1.12"]
build-backend = "poetry.masonry.api"
